+++ packages/react-ui/src/app/mcp/mcp-piece.tsx
@@ -1,23 +1,27 @@
-import { McpPieceStatus, McpPieceWithConnection } from '@activepieces/shared';
+import { useMutation } from '@tanstack/react-query';
 import { t } from 'i18next';
 import { Pencil, PlugIcon, Trash2 } from 'lucide-react';
-
-import { Card, CardContent } from '../../../components/ui/card';
-import { PieceIcon } from '../../pieces/components/piece-icon';
+import { useState } from 'react';
 
 import { ConfirmationDeleteDialog } from '@/components/delete-dialog';
-import { Skeleton } from '@/components/ui/skeleton';
-import { McpPieceDialog } from './mcp-piece-dialog';
 import { Button } from '@/components/ui/button';
-import { TooltipContent, TooltipTrigger,Tooltip } from '@/components/ui/tooltip';
-import { piecesHooks } from '@/features/pieces/lib/pieces-hook';
-import { useMutation } from '@tanstack/react-query';
-import { mcpApi } from '../lib/mcp-api';
-import { toast } from '@/components/ui/use-toast';
-import { Toggle } from '@/components/ui/toggle';
+import { Skeleton } from '@/components/ui/skeleton';
 import { Switch } from '@/components/ui/switch';
-import { mcpHooks } from '../lib/mcp-hooks';
-import { useState } from 'react';
+import {
+  TooltipContent,
+  TooltipTrigger,
+  Tooltip,
+} from '@/components/ui/tooltip';
+import { toast } from '@/components/ui/use-toast';
+import { piecesHooks } from '@/features/pieces/lib/pieces-hook';
+import { McpPieceStatus, McpPieceWithConnection } from '@activepieces/shared';
+
+import { Card, CardContent } from '../../components/ui/card';
+import { mcpApi } from '../../features/mcp/lib/mcp-api';
+import { mcpHooks } from '../../features/mcp/lib/mcp-hooks';
+import { PieceIcon } from '../../features/pieces/components/piece-icon';
+
+import { McpPieceDialog } from './mcp-piece-dialog';
 
 type McpPieceProps = {
   piece: McpPieceWithConnection;
@@ -35,12 +39,12 @@ export const McpPiece = ({
   onDelete,
   isLoading = false,
 }: McpPieceProps) => {
-  const {pieceModel, isLoading: isPieceLoading} = piecesHooks.usePiece({
+  const { pieceModel, isLoading: isPieceLoading } = piecesHooks.usePiece({
     name: piece.pieceName,
   });
 
-  const {refetch: refetchMcp} = mcpHooks.useMcp();
-  const {mutate: updatePieceStatus} = useMutation({
+  const { refetch: refetchMcp } = mcpHooks.useMcp();
+  const { mutate: updatePieceStatus } = useMutation({
     mutationFn: async (status: McpPieceStatus) => {
       setStatus(status === McpPieceStatus.ENABLED);
       await mcpApi.updatePiece({
@@ -52,7 +56,11 @@ export const McpPiece = ({
     onSuccess: (status) => {
       toast({
         title: `${t('MCP tool')} (${pieceModel?.displayName})`,
-        description: t(`${status === McpPieceStatus.ENABLED ? 'Enabled' : 'Disabled'} successfully`),
+        description: t(
+          `${
+            status === McpPieceStatus.ENABLED ? 'Enabled' : 'Disabled'
+          } successfully`,
+        ),
       });
       refetchMcp();
     },
@@ -62,7 +70,7 @@ export const McpPiece = ({
         description: t('Failed to update piece status'),
         variant: 'destructive',
       });
-    }
+    },
   });
   const [status, setStatus] = useState(piece.status === McpPieceStatus.ENABLED);
   if (isLoading || isPieceLoading) {
@@ -83,16 +91,13 @@ export const McpPiece = ({
       </Card>
     );
   }
-  
+
   const displayName = piece.connection
     ? piece.connection.displayName
     : pieceInfo.displayName;
-  
+
   return (
     <Card className="overflow-hidden transition-all duration-200 relative hover:shadow-sm group border-border min-w-[420px]">
-       
-      
-
       <CardContent className="flex flex-row items-center p-4 justify-between ">
         <div className="flex items-center space-x-3 min-w-0 py-1.5">
           <div className="relative">
@@ -112,83 +117,63 @@ export const McpPiece = ({
           </div>
           <div className="min-w-0">
             <h4 className="font-medium text-foreground truncate flex items-center gap-1">
-            {pieceInfo.displayName}
+              {pieceInfo.displayName}
             </h4>
-            {  piece.connection && 
+            {piece.connection && (
               <p className="text-xs text-muted-foreground mt-0.5">
-              {displayName}
-                
+                {displayName}
               </p>
-            }
-           
+            )}
           </div>
-        
         </div>
         <div className="flex items-center gap-2">
+          <Switch
+            checked={status}
+            onCheckedChange={(checked) => {
+              updatePieceStatus(
+                checked ? McpPieceStatus.ENABLED : McpPieceStatus.DISABLED,
+              );
+            }}
+          ></Switch>
 
-
-
-        <Switch
-      checked={status}
-      onCheckedChange={(checked) => {
-        updatePieceStatus(checked ? McpPieceStatus.ENABLED : McpPieceStatus.DISABLED);
-      }}
-      
-      >
-
-      </Switch>
-
-{
-  pieceModel?.auth && (  
-  <Tooltip>
-    <McpPieceDialog mcpPieceToUpdate={piece}>
-    <TooltipTrigger asChild>
-      <Button variant='ghost' size='icon'>
-        <Pencil className="h-4 w-4  " />
-      </Button>
-  </TooltipTrigger>
-
-    </McpPieceDialog>
-  <TooltipContent>
-    {t('Edit Connection')}
-  </TooltipContent>
-</Tooltip>
-)
-}
-
-     
-          
-
+          {pieceModel?.auth && (
+            <Tooltip>
+              <McpPieceDialog mcpPieceToUpdate={piece}>
+                <TooltipTrigger asChild>
+                  <Button variant="ghost" size="icon">
+                    <Pencil className="h-4 w-4  " />
+                  </Button>
+                </TooltipTrigger>
+              </McpPieceDialog>
+              <TooltipContent>{t('Edit Connection')}</TooltipContent>
+            </Tooltip>
+          )}
 
           <Tooltip>
-          <ConfirmationDeleteDialog
-        title={`${t('Delete')} ${displayName}`}
-        message={<div>{t('Are you sure you want to delete this tool from your MCP? if you delete it you won\'t be able to use it in your MCP client.')}</div>}
-        mutationFn={async () => {
-          onDelete(piece);
-        }}
-        entityName={t('piece')}
-      >
-          <TooltipTrigger asChild>
-        <Button variant='ghost' size='icon'>
-        <Trash2 className="h-4 w-4 text-destructive transition-colors duration-200 group-hover:text-destructive/90" />
+            <ConfirmationDeleteDialog
+              title={`${t('Delete')} ${displayName}`}
+              message={
+                <div>
+                  {t(
+                    "Are you sure you want to delete this tool from your MCP? if you delete it you won't be able to use it in your MCP client.",
+                  )}
+                </div>
+              }
+              mutationFn={async () => {
+                onDelete(piece);
+              }}
+              entityName={t('piece')}
+            >
+              <TooltipTrigger asChild>
+                <Button variant="ghost" size="icon">
+                  <Trash2 className="h-4 w-4 text-destructive transition-colors duration-200 group-hover:text-destructive/90" />
+                </Button>
+              </TooltipTrigger>
+            </ConfirmationDeleteDialog>
 
-        </Button>
-        </TooltipTrigger>
-        
-      </ConfirmationDeleteDialog>
-       
-        <TooltipContent>
-          {t('Delete')}
-        </TooltipContent>
-        </Tooltip>
-
-    
-      
-
-  
-      
-          </div>
+            <TooltipContent>{t('Delete')}</TooltipContent>
+          </Tooltip>
+        </div>
       </CardContent>
     </Card>
   );
