+++ packages/server/api/src/app/mcp/mcp-module.ts
@@ -1,8 +1,8 @@
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox'
 import { mcpPieceController } from './mcp-piece-controller'
+import { mcpServerController } from './mcp-server-controller'
 import { mcpSessionManager } from './mcp-session-manager'
 import { mcpSseController } from './mcp-sse-controller'
-import { mcpServerController } from './mcp-server-controller'
 
 export const mcpModule: FastifyPluginAsyncTypebox = async (app) => {
     await app.register(mcpServerController, { prefix: '/v1/mcp-servers' })
