+++ packages/server/shared/package.json
@@ -4,7 +4,7 @@
   "dependencies": {
     "tslib": "^2.3.0",
     "@sentry/node": "7.120.0",
-    "pino": "9.5.0",
+    "pino": "9.6.0",
     "pino-loki": "2.1.3",
     "@activepieces/shared": "*",
     "async-mutex": "0.4.0",
