+++ packages/server/worker/package.json
@@ -7,7 +7,7 @@
   "dependencies": {
     "@activepieces/pieces-framework": "0.7.46",
     "@activepieces/server-shared": "0.0.1",
-    "@activepieces/shared": "0.10.170",
+    "@activepieces/shared": "0.10.171",
     "async-mutex": "0.4.0",
     "axios": "1.8.3",
     "axios-retry": "4.4.1",
