+++ packages/shared/package.json
@@ -1,5 +1,5 @@
 {
   "name": "@activepieces/shared",
-  "version": "0.10.170",
+  "version": "0.10.171",
   "type": "commonjs"
 }
